Segregate controller for Jwks management
