+++ controller/Jwks.php
@@ -24,16 +24,34 @@ namespace oat\taoLti\controller;
 
 use oat\tao\helpers\UrlHelper;
 use oat\tao\model\security\Business\Contract\JwksRepositoryInterface;
+use oat\taoLti\models\classes\Platform\Service\CachedKeyChainGenerator;
 use oat\taoLti\models\classes\Security\DataAccess\Repository\CachedPlatformJwksRepository;
-use \tao_actions_CommonModule;
+use \tao_actions_CommonModule as CommonModule;
 
-class Jwks extends tao_actions_CommonModule
+class Jwks extends CommonModule
 {
-    public function index(): void
+    public function view(): void
     {
         $this->setData('jwks-key', json_encode($this->getJwksRepository()->find()));
         $this->setData('jwks-generate-url', $this->getUrlGenerator()->buildUrl('jwks', 'Security'));
         $this->setView('jwks/Jwks.tpl');
+    }
+
+    public function index(): void
+    {
+        switch ($this->getRequestMethod()) {
+            case 'POST':
+                $this->postJwks();
+                break;
+
+            default:
+                $this->setResponse($this->getPsrResponse()->withStatus(501));
+        }
+    }
+
+    private function postJwks(): void
+    {
+        $this->getCachedKeyChainGenerator()->generate();
     }
 
     private function getJwksRepository(): JwksRepositoryInterface
@@ -44,5 +62,10 @@ class Jwks extends tao_actions_CommonModule
     private function getUrlGenerator(): UrlHelper
     {
         return $this->getServiceLocator()->get(UrlHelper::class);
+    }
+
+    private function getCachedKeyChainGenerator(): CachedKeyChainGenerator
+    {
+        return $this->getServiceLocator()->get(CachedKeyChainGenerator::class);
     }
 }
