+++ doc/rest.json
@@ -366,18 +366,23 @@
                         "description": "Internal error (should not occur)"
                     }
                 }
-            },
+            }
+        },
+        "/taoLti/Jwks/index": {
             "post": {
                 "description": "Generate the public JWKS used on LTI 1.3 communication",
                 "tags": [
-                    "security"
+                    "security", "jwks"
                 ],
                 "responses": {
                     "200": {
-                        "description": "Acknowledge JWKS generation has been successful",
+                        "description": "Acknowledge JWKS generation has been successful"
                     },
                     "500": {
                         "description": "Internal error (should not occur)"
+                    },
+                    "501": {
+                        "description": "Internal error (Http method not implemented)"
                     }
                 }
             }
